+++ packages/fsengagement/src/types.ts
@@ -8,24 +8,23 @@ import {
   ViewStyle
 } from 'react-native';
 import { Navigator } from '@brandingbrand/fsapp';
+import { EngagementService } from './EngagementService';
 
 export interface ScreenProps {
   componentId: string;
 }
 
-export interface Action {
+export interface EmitterProps {
+  id?: string;
+  name?: string;
+}
+
+export interface Action extends EmitterProps {
   type: string;
   value: string;
   subject?: string;
   body?: string;
-  name?: string;
-  id?: string;
   position?: number;
-}
-
-export interface EmitterProps {
-  id?: string;
-  name?: string;
 }
 
 export interface ComponentList {
@@ -41,16 +40,14 @@ export interface Icon {
   iconStyle?: StyleProp<ImageStyle>;
 }
 
-export interface CardProps {
+export interface CardProps extends EmitterProps {
   containerStyle?: StyleProp<TextStyle>;
   private_blocks: BlockItem[];
   story?: JSON;
-  api?: any;
+  api?: EngagementService;
   plainCard?: boolean;
   storyGradient?: StoryGradient;
   navigator: Navigator;
-  name?: string;
-  id?: string;
 }
 
 export interface Empty {
@@ -72,7 +69,7 @@ export interface HTML {
   title: JSON;
 }
 
-export interface JSON {
+export interface JSON extends EmitterProps {
   isBlog?: boolean;
   backArrow?: StyleProp<ImageStyle>;
   private_blocks?: BlockItem[];
@@ -87,7 +84,6 @@ export interface JSON {
   containerStyle?: StyleProp<ViewStyle>;
   id?: string;
   key?: string;
-  name?: string;
   storyType?: string;
   tabbedItems?: any[];
   AnimatedPageCounter?: any;